refactor: moved set_indentation to utils module
