+++ backend/src/lib.rs
@@ -3,3 +3,4 @@ pub mod configuration;
 pub mod doc;
 pub mod errors;
 pub mod execution;
+pub mod utils;
